refactor: create separate classes for test constant and test data factories
